Added RGB range checks for obstacles and rock samples
Proved that it works for the actual rover images.

For Obstacles, I used lower and upper bounds from a range of brownish RGBs,
and that seems to work to isolate obstacles (rocks and mountains).

For rock samples, the above trick did not work. So I used EOG's mouse over to
visually inspect the RGB values and put those in the range. Initially it included
the rock sample with some landscape silhoutte area around it, but I could isolate
the rock itself very accurately by lowering down the Z (3rd channel's value) to a
range that seems to work..