Reduce time to detect stuck, and steer more towards left wall now
Still imperfect and misses some nooks (where a jewel may be kept).
But the unstuck code kicks in more often (if stuck towards a wall)

More often than not the simulator's code eventually unstucks the rover
(the control code), but there is a place (towards left of global map)
where the rover got really stuck.

At such time, it needs to be detected, and steering direction can be double
applied .. the question is how to detect that "we are still stuck during unstuck mode",
i.e. not making sufficient progress.