Increase rock collection probability
When near sample, slow down the rover. There are a few
cases where it overshoots the sample so need to check.

Remove some extra printfs

The next part now is to steer the rover towards the sample
when it 'locates' it. Use code similar to Rock_WorldMap usage
in supporting functions, and combine it with Rover's current
position to manipulate the steer (and not override it whilst
in the goal reaching state)