Add heuristics to calculate absolute distance between rover and rock
Discover a new rock near trajectory at a distance of 10 (x or y).
Use absolute instead of squareroot.

Also my rock detection color scheme is awry ! It DOES NOT work with
a much better GPU resolution :( In fact at setting set to "good" or
"fantastic", at every movement, a few pixels are incorrectly classified
as rock samples).The 10 distance somewhat acts as a filter but this problem
needs to be addressed.