Display rover's coordinates on vision map
Add code to start changing speed and steer of rover based on upcoming rock
and based on direction.

While structurally it makes sense, it does not work. This could be because
the logic is too simple, or some variables are overriden by sim (vel?) or I
am simply using incorrect values, or the relative ranges of some variable values
are not normalized to each other, or other variables like brake or throttle have to
be deployed, or that values by which steer is changed is not enough (+12 and above)

All that has to be explored next