The robot tracks L, C and R, and catches the ball !
Got it accurately in the first shot itself.

We need to convert 2D to 1D. I took the example of an 8 by 8 matrix
and realized that the array is | L-range | C-range | R-range | L-range | ... repeat.

To determine these ranges accurately, the current row must be known.
Step, is so called, because it helps determine exactly that, in this given 1D array

1/3d is 0.34 and 3/2rd is 0.67

Thus:
L-range is ( 0 to step * 0.34 ) normalized to the current row (i.e. mult by step_size * current_row)
and similarily for the other 2 ranges.

Also when the white pixel is towards the extreme left upper quadrant, I assume we are very close
to the ball and I order the robot to stop.

Works fine - I tried all of Right, Center and Left cases.