Make chassis a square instead of rect
Also add a second .xacro robot model that more resembles an R2D2 shape ..

Checked that with the squared chassis, my robot still chases the ball.
Only thing is that its not in view now, so jump to top view to make it work.

Can also save the world file with the correct orientation, but not now ..