Integrate rviz with the 4 packages
Edit the rviz configuration:
Change the Fixed Frame to base_footprint
Change the Reference Frame to odom
Add a RobotModel
Add a camera and select the /camera/rgb/image_raw topic
Add a /ekfpath topic and change the display name to EKFPath
Add a /odompath topic and change the display name to OdomPath
Change the OdomPath color to red:255;0;0

Checked that this works. Using teleop package, I move the robot
and can see how the ekfpath is drawn accurately over time.