Compute particle weights, based on a probability of goodness of the landmark measures
The goodness function again takes the 8 landmark distance values (observed by the actual robot)
and compares (recomputes) the distance of the real landmarks with the x,y of the fake particle,
and accumulates all 8 (multiplying) into a single probability value, that also takes in gaussian
noise data into account. This single value goodness-probability becomes the weight of that particle.

Thus the weight is based on the actual computation (z) compared with the x,y of the particle, returned
as a goodness of measure probability.