Add teleop package and modify configuration files
Added and tested (that it launches) the optional teleop kbd package
just because I feel it may be easier to control the movement of the robot
when checking AMCL

Since I was recieving warnings on Transform_Timeout and Map Update Loop, I
reconfigured the parameters resptvly, i.e. larger transform_tolerance for the
AMCL node and lower update_drequency and publish_drequency values in the config files.
(It did not help though)