Add a gen-map script for pgm_maps (home)
I have an error on some machines with libProtobuf saying that
vector2d.proto already exists in the database. I tried commenting out
the relevant msg generation lines in the cmakefile of msgs, but then I
get the error that file vector2d.pb.h is not found.

Anyways, as the main packages still work (and map gen works on a diff machine)
we can move ahead for now ..