Fixed the laser and transform errors in rviz! (incorrect rescan in amcl launch file)
The laser scan topic is actually /scan and not my_robot/laser/scan currently
(contrary to what the course suggests). This I verified by 'rostopic hz /scan'
which spit out data. Also I read up that amcl node may not even start computing till
it recieved correct scan data.

So I have removed the rescanning part in the amcl launch file, for both amcl and
move-base node. The result is that as soon as rviz gets the 3000 * 3000 map data
from map-server, the robot is correctly placed inside the map (the transform code
for mapping executes, plus the red-arrows are generated, meaning that localization
algo is now running) - not a whole lot of warning msgs remain!

Upon moving the robot by teleop, I see that a large splash or red surrounds the robot,
meaning that amcl optimizations still have to be done. One thing is that perhaps the landmarks
in the room (detected by the laser, that's why we need lasers, to look outside walls!) may
not be enough? Will try placing the robot in different positions as well ..