Fixed subscribing to /odom issue
Saw that there are 2 odom topics /odom and /rtabmap/odom.
Subscribe to /odom, the easiest way is to remove the "group ns" setting
in mapping.launch, which seems to have everything subscribe to rtabmap

No more (no data after 5 sec) errors. And the odometry window shows up.

What I noticed is that the odom window contains the same image data! (no depth
data). Also the loop closure window is black. Now that could be because my
compute here is horribly slow. So I will have to test this on a faster machine
(and compare the output to the previous output where I saw those 'data bubbles/flares')

That apart, when I used loop closure constraint as ICP instead of visual (captured
in mapping.launch.OLD), I still saw loop closure rejects but the reason this time is
that some hessians are incorrect (plus the cloud map is all NaNs). Needs debugging.