Add gmapping + turtlebot_gazebo solution under samples/gmapping
Tested that it works by running SLAM.sh (turtlebot moves with commands and
gmapping generates map seen in rviz)

Also in task5, compile in the other packages and launch. The only thing
here is that task5's turtlebot doesnt move, and I guess this is because
of the double teleop package present there.

This big commit checks in the required devel and build files for both
samples/gmapping and for task5.