Add amcl.launch file from task3 to this project
Using it as such led to a couple of errors (which I checked only once)
so I changed robot_footprint back to base_footprint. (This may need to
be verified).

The real map gets pulled in because myrobot package still exists here
and the map is pulled from that map folder. The red arrows still don't
show up here.

Current issues are:
1. turtlebot_gazebo does not seem to get localized (or too well?) because I dont
see any red arrows
2. move_base does not seem to work