Pass actual map of my_world in test_navigation.
This is because amcl will compare the laser scan to the actual map
and then localize. I also see that after some unsuccessful paths
(mostly because of the structure of the map - its not got any doors
or windows yet), if the end goal is within the same room and amcl/move_base
thinks robot is free to move, it will move it to its eventual goal

The screenshot attached shows how this works.