Establish pick and drop zone topics for home_service to be more robust
The current odometry data is not reliable (to be debugged). In the interim,
I have designed an alternate, easier approach where pick_object will publish
simple topic when the robot reaches either the pick zone or the drop zone.

Subscribers like the home_service upon listening can do their marker management
(in this case either deleting + waiting for 5 sec, or showing an object dropped-off)

Tested that this works. Once the robot reaches the pick-up zone, the marker is hidden
for 5 sec, and one it goes to the drop zone, the marker is published again.